Add logout & account handlers and remove some comments NOTE: 1. logoutHandler() function in logout.handler.js 2. accountHandler() function in account.handler.js 3. sampleProducts variable in sample.products.js
