+++ frontend/JS/auth.js
@@ -1,37 +1,5 @@
-import applyTheme from "./theme.js";
-
-const sampleProducts = [
-  {
-    id: 1,
-    title: "قاب گوشی اپل",
-    price: "120,000 تومان",
-    image: "/src/images/1.jpg",
-  },
-  {
-    id: 2,
-    title: "قاب گوشی سامسونگ",
-    price: "150,000 تومان",
-    image: "/src/images/2.jpg",
-  },
-  {
-    id: 3,
-    title: "قاب گوشی شیایومی",
-    price: "130,000 تومان",
-    image: "/src/images/3.jpg",
-  },
-  {
-    id: 4,
-    title: "قاب گوشی هواوی",
-    price: "200,000 تومان",
-    image: "/src/images/4.jpg",
-  },
-  {
-    id: 5,
-    title: "قاب گوشی ال جی",
-    price: "200,000 تومان",
-    image: "/src/images/5.jpg",
-  },
-];
+import applyTheme from "./utils/theme.js";
+import sampleProducts from "./utils/sample.products.js";
 
 const BASE_URL = "http://localhost:3000";
 const loginForm = document.getElementById("login-form");
