+++ frontend/JS/utils/account.handler.js
@@ -1,0 +1,11 @@
+function accountHandler() {
+  const accountBtn = document.getElementById("account-btn");
+  if (accountBtn) {
+    accountBtn.addEventListener("click", (e) => {
+      e.preventDefault();
+      window.location.href = "profile.html";
+    });
+  }
+}
+
+export default accountHandler;
